requirements.txt
# Core MCP Framework
mcp==0.1.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Document Processing
docling==1.0.0
python-docx==0.8.11
pdf2image==1.16.3
Pillow==10.1.0

# Vector Database & Search
chromadb==0.4.24
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Database & Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# API & Web Integration
requests==2.31.0
aiohttp==3.9.1
google-generativeai==0.3.0

# Async & Task Processing
celery==5.3.4
redis==5.0.1
asyncio==3.4.3

# Caching & Performance
cachetools==5.3.2
diskcache==5.6.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.0
pylint==3.0.3
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
python-json-logger==2.0.7
prometheus-client==0.19.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Utilities
python-multipart==0.0.6
email-validator==2.1.0
